IPRT/time: Fix handling of negative UTC offset in the POSIX variant of RTTimeLocalExplode (type propagation made it unsigned).
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107965 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/posix/timelocal-posix.cpp
@@ -208,7 +208,7 @@ RTDECL(PRTTIME) RTTimeLocalExplode(PRTTIME pTime, PCRTTIMESPEC pTimeSpec)
     if (pTime)
     {
         pTime->fFlags = (pTime->fFlags & ~RTTIME_FLAGS_TYPE_MASK) | RTTIME_FLAGS_TYPE_LOCAL;
-        pTime->offUTC = cNsUtcOffset / RT_NS_1MIN;
+        pTime->offUTC = cNsUtcOffset / (int64_t)RT_NS_1MIN;
     }
     return pTime;
 }
